module-04-incident-response-operational-excellence/demo-clip4-blameless-postmortems-mttr-backlog/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
prometheus-client==0.20.0
pydantic==2.8.2
pyyaml==6.0.1
